trac-1048. Make size fields be 64bit. Sysfs now reports everything in bytes. Control create accepts 64bit, byte integer as well.

+++ kernel/castle_public.h
@@ -364,7 +364,7 @@ struct castle_slave_superblock_public {
     uint32_t version;   /* Super chunk format version */
     uint32_t uuid;
     uint32_t used;
-    uint32_t size; /* In blocks */
+    uint64_t size;      /* In blocks. */
 	uint32_t flags;
 } PACKED;
